refactoring folder structure and files in test folder
