+++ tests/config/webhooks.test.js
@@ -1,0 +1,1 @@
+// Tests
