+++ tests/utils/formattingUtils.test.js
@@ -1,15 +1,18 @@
-import { CHANGELOG_ENTRY_PREFIXES, MAX_ENTRY_LENGTH } from "../config/constants.js";
+import {
+  CHANGELOG_ENTRY_PREFIXES,
+  MAX_ENTRY_LENGTH,
+} from "../../config/constants.js";
 import {
   prepareChangelogEntry,
   prepareChangelogEntriesMap,
   prepareChangesetEntriesContent,
-} from "../utils/formattingUtils.js";
+} from "../../utils/formattingUtils.js";
 import {
   InvalidPrefixError,
   EmptyEntryDescriptionError,
   EntryTooLongError,
   ChangelogEntryMissingHyphenError,
-} from "../utils/customErrors.js";
+} from "../../utils/customErrors.js";
 
 describe("Formatting Utils Tests", () => {
   const prNumber = 123;
@@ -68,7 +71,7 @@ describe("Formatting Utils Tests", () => {
 
     test("with entry too long, should throw EntryTooLongError", () => {
       const longEntryText =
-        " a very long entry with too much text that exceeds the maximum allowed length limit specified for for a changelog entry"
+        " a very long entry with too much text that exceeds the maximum allowed length limit specified for for a changelog entry";
       const longEntry = `- feat:${longEntryText}`;
       const characterOverage = longEntryText.length - MAX_ENTRY_LENGTH;
       const expectedErrorMessage = `Entry is ${
